Спринт 19.3 SettingsFragment.kt  подключен, активити удалена.
