+++ app/src/main/java/com/practicum/playlistmaker/settings/ui/SettingsFragment.kt
@@ -1,60 +1,71 @@
 package com.practicum.playlistmaker.settings.ui
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.practicum.playlistmaker.R
+import androidx.core.net.toUri
+import com.practicum.playlistmaker.databinding.FragmentSettingsBinding
+import org.koin.androidx.viewmodel.ext.android.viewModel
+import kotlin.getValue
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [SettingsFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class SettingsFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+    private val viewModel: SettingsViewModel by viewModel()
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
+    private lateinit var binding: FragmentSettingsBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_settings, container, false)
+    ): View {
+        binding = FragmentSettingsBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment SettingsFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            SettingsFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        viewModel.observeTheme()
+            .observe(viewLifecycleOwner) { data ->
+                binding.themeSwitcher.isChecked = data
             }
+
+        binding.themeSwitcher.setOnCheckedChangeListener { _, checked ->
+            viewModel.changeDarkTheme(checked)
+        }
+
+        binding.shareLabel.setOnClickListener {
+            val data = viewModel.shareApp()
+            val intent = Intent(Intent.ACTION_SEND)
+            val chooserTitle = data.title
+            intent.type = "text/plain"
+            intent.putExtra(
+                Intent.EXTRA_TEXT,
+                data.text
+            )
+            startActivity(Intent.createChooser(intent, chooserTitle))
+        }
+
+        binding.supportLabel.setOnClickListener {
+            val data = viewModel.writeToSupport()
+            val intent = Intent(Intent.ACTION_SENDTO)
+            val chooserTitle = data.title
+
+            val mailto = "mailto:${data.address}" +
+                    "?subject=${Uri.encode(data.subj)}" +
+                    "&body=${Uri.encode(data.body)}"
+
+            intent.data = mailto.toUri()
+            startActivity(Intent.createChooser(intent, chooserTitle))
+        }
+
+        binding.agreementLabel.setOnClickListener {
+            val data = viewModel.userAgreement()
+            val intent = Intent(Intent.ACTION_VIEW, data.value.toUri())
+            startActivity(intent)
+        }
     }
 }
